On FreeBSD 12+ vmmeter.v_cache_count is removed.
See:

https://reviews.freebsd.org/D8583
https://svnweb.freebsd.org/base?view=revision&revision=309017

Dummy values (always returning 0) were added if COMPAT_FREEBSD11 is
defined. This is defined in GENERIC, so will be there for most users.

This fixes #103 by testing for vmmeter.v_cache_count and setting the
cache value to 0 if it doesn't exist. However, more investigation may
need to be done to see if cache can be obtained in a different way.